move the audio count to a server-side API call
This moves the audio count from a static value that I had to manually
set each time I updated the number of files, to an API call. This means
I just need to upload the files and mayyyybe invalidate the cache, as
opposed to having to update the front end and whatnot.

There's a local fallback to 43 included as well.